requirements.txt
Django>=4.2
djangorestframework
requests
python-dotenv
chromadb
openai   # or your preferred LLM provider
